refactor[#279] : dashboard -> getProfile  파일명 변경
