+++ src/entities/profile/api/index.ts
@@ -1,1 +1,1 @@
-export * from "./dashboard";
+export * from "./getProfile";
